refactor: move retry annotations and extensions to main
