+++ build.gradle
@@ -65,7 +65,7 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter'
 
     testImplementation 'io.qameta.allure:allure-junit5:2.29.0'
-    testImplementation 'io.github.artsok:rerunner-jupiter:2.1.6'
+    implementation 'io.github.artsok:rerunner-jupiter:2.1.6'
     testRuntimeOnly 'org.aspectj:aspectjweaver:1.9.22.1'
 }
